Rename UML components package to deployments
