+++ gaphor/SysML/blocks/connectors.py
@@ -8,7 +8,7 @@ from gaphor.SysML import sysml
 from gaphor.SysML.blocks.block import BlockItem
 from gaphor.SysML.blocks.property import PropertyItem
 from gaphor.SysML.blocks.proxyport import ProxyPortItem
-from gaphor.UML.components import ConnectorItem
+from gaphor.UML.deployments import ConnectorItem
 
 
 @Connector.register(BlockItem, ProxyPortItem)
