+++ gaphor/SysML/blocks/tests/test_connectors.py
@@ -7,7 +7,7 @@ from gaphor.SysML import sysml
 from gaphor.SysML.blocks.block import BlockItem
 from gaphor.SysML.blocks.connectors import BlockProperyProxyPortConnector
 from gaphor.SysML.blocks.proxyport import ProxyPortItem
-from gaphor.UML.components import ConnectorItem
+from gaphor.UML.deployments import ConnectorItem
 
 
 @pytest.fixture
