--- gaphor/UML/components/__init__.py
@@ -1,12 +1,0 @@
-from gaphor.UML.components import (
-    componentsgrouping,
-    componentspropertypage,
-    connectorconnect,
-    copypaste,
-)
-from gaphor.UML.components.artifact import ArtifactItem
-from gaphor.UML.components.component import ComponentItem
-from gaphor.UML.components.connector import ConnectorItem
-from gaphor.UML.components.node import NodeItem
-
-__all__ = ["ArtifactItem", "ComponentItem", "ConnectorItem", "NodeItem"]
