+++ gaphor/UML/deployments/__init__.py
@@ -1,0 +1,12 @@
+from gaphor.UML.deployments import (
+    componentsgrouping,
+    componentspropertypage,
+    connectorconnect,
+    copypaste,
+)
+from gaphor.UML.deployments.artifact import ArtifactItem
+from gaphor.UML.deployments.component import ComponentItem
+from gaphor.UML.deployments.connector import ConnectorItem
+from gaphor.UML.deployments.node import NodeItem
+
+__all__ = ["ArtifactItem", "ComponentItem", "ConnectorItem", "NodeItem"]
