+++ gaphor/UML/deployments/artifact.py
