+++ gaphor/UML/deployments/component.py
