+++ gaphor/UML/deployments/copypaste.py
