+++ gaphor/UML/deployments/tests/test_copypaste.py
@@ -8,7 +8,7 @@ from gaphor.core.modeling.modelinglanguage import (
 from gaphor.diagram.tests.fixtures import connect, copy_clear_and_paste_link
 from gaphor.SysML.modelinglanguage import SysMLModelingLanguage
 from gaphor.UML.classes import InterfaceItem
-from gaphor.UML.components import ComponentItem, ConnectorItem
+from gaphor.UML.deployments import ComponentItem, ConnectorItem
 from gaphor.UML.modelinglanguage import UMLModelingLanguage
