+++ gaphor/UML/diagramitems.py
@@ -6,7 +6,7 @@ This module makes it easier to load a diagram item.
 from gaphor.diagram.general import *
 from gaphor.UML.actions import *
 from gaphor.UML.classes import *
-from gaphor.UML.components import *
+from gaphor.UML.deployments import *
 from gaphor.UML.interactions import *
 from gaphor.UML.profiles import *
 from gaphor.UML.states import *
