+++ gaphor/diagram/tools/tests/test_dropzone.py
@@ -2,7 +2,7 @@ from gaphas.item import NW, SE
 
 from gaphor import UML
 from gaphor.diagram.tools.dropzone import drop_zone_tool, grow_parent, on_motion
-from gaphor.UML.components import ComponentItem, NodeItem
+from gaphor.UML.deployments import ComponentItem, NodeItem
 
 
 def test_hover_over_drop_zone(diagram, element_factory, view):
